requirements.txt
asgiref==3.6.0
Django==4.1.7
django-unique-slugify==1.1
djangorestframework==3.14.0
psycopg2==2.9.5
python-dotenv==0.21.1
pytils==0.4.1
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
